Moved fixture docker_client_config to conftest.
